+++ src/Hmac/HmacCsrfToken.php
@@ -4,36 +4,39 @@ declare(strict_types=1);
 
 namespace Yiisoft\Csrf\Hmac;
 
+use RuntimeException;
 use Yiisoft\Csrf\CsrfTokenInterface;
 use Yiisoft\Csrf\Hmac\IdentityGenerator\CsrfTokenIdentityGeneratorInterface;
-use Yiisoft\Security\DataIsTamperedException;
-use Yiisoft\Security\Mac;
-use Yiisoft\Strings\StringHelper;
 use Yiisoft\Csrf\MaskedCsrfToken;
+use Yiisoft\Security\Random;
+use Yiisoft\Strings\StringHelper;
 
 use function count;
+use function hash_equals;
+use function hash_hmac;
 
 /**
- * Stateless CSRF token that does not require any storage. The token is a hash from session ID and a timestamp
- * (to prevent replay attacks). It is added to forms. When the form is submitted, we re-generate the token from
- * the current session ID and a timestamp from the original token. If two hashes match, we check that timestamp is
- * less than {@see HmacCsrfToken::$lifetime}.
- *
- * The algorithm is also known as "HMAC Based Token".
+ * Stateless CSRF token that does not require any storage. The token contains expiration timestamp and random value,
+ * and is signed with a session-bound identity. It is added to forms. When the form is submitted, we verify the token
+ * signature, check that it belongs to the current session identity, and check that it has not expired.
  *
  * Do not forget to decorate the token with {@see MaskedCsrfToken} to prevent BREACH attack.
  *
- * @link https://cheatsheetseries.owasp.org/cheatsheets/Cross-Site_Request_Forgery_Prevention_Cheat_Sheet.html#hmac-based-token-pattern
+ * @link https://cheatsheetseries.owasp.org/cheatsheets/Cross-Site_Request_Forgery_Prevention_Cheat_Sheet.html#employing-hmac-csrf-tokens
  */
 final class HmacCsrfToken implements CsrfTokenInterface
 {
     private CsrfTokenIdentityGeneratorInterface $identityGenerator;
-    private Mac $mac;
 
     /**
      * @var string Shared secret key used to generate the hash.
      */
     private string $secretKey;
+
+    /**
+     * @var string Hash algorithm for message authentication.
+     */
+    private string $algorithm;
 
     /**
      * @var int|null Number of seconds that the token is valid for.
@@ -47,8 +50,8 @@ final class HmacCsrfToken implements CsrfTokenInterface
         ?int $lifetime = null
     ) {
         $this->identityGenerator = $identityGenerator;
-        $this->mac = new Mac($algorithm);
         $this->secretKey = $secretKey;
+        $this->algorithm = $algorithm;
         $this->lifetime = $lifetime;
     }
 
@@ -66,39 +69,43 @@ final class HmacCsrfToken implements CsrfTokenInterface
             return false;
         }
 
-        [$expiration, $identity] = $data;
+        [$expiration, $payload] = $data;
+
+        $hashLength = $this->getHashLength();
+        $hash = StringHelper::byteSubstring($payload, 0, $hashLength);
+        $message = StringHelper::byteSubstring($payload, $hashLength, null);
+
+        if (!hash_equals($hash, $this->generateHash($message))) {
+            return false;
+        }
 
         if ($expiration !== null && time() > $expiration) {
             return false;
         }
-
-        return $identity === $this->identityGenerator->generate();
+        return true;
     }
 
     private function generateToken(?int $expiration): string
     {
-        return StringHelper::base64UrlEncode(
-            $this->mac->sign(
-                (string) $expiration . '~' . $this->identityGenerator->generate(),
-                $this->secretKey,
-                true,
-            ),
-        );
+        $message = (string) $expiration . '~' . Random::string(32);
+
+        return StringHelper::base64UrlEncode($this->generateHash($message) . $message);
     }
 
+    /**
+     * @return array{0: int|null, 1: string}|null
+     */
     private function extractData(string $token): ?array
     {
-        try {
-            $raw = $this->mac->getMessage(
-                StringHelper::base64UrlDecode($token),
-                $this->secretKey,
-                true,
-            );
-        } catch (DataIsTamperedException $e) {
+        $payload = StringHelper::base64UrlDecode($token);
+        $hashLength = $this->getHashLength();
+
+        if (StringHelper::byteLength($payload) <= $hashLength) {
             return null;
         }
 
-        $chunks = explode('~', $raw, 2);
+        $message = StringHelper::byteSubstring($payload, $hashLength, null);
+        $chunks = explode('~', $message, 2);
         if (count($chunks) !== 2) {
             return null;
         }
@@ -112,8 +119,22 @@ final class HmacCsrfToken implements CsrfTokenInterface
             }
         }
 
-        $identity = $chunks[1];
+        return [$expiration, $payload];
+    }
 
-        return [$expiration, $identity];
+    private function generateHash(string $message): string
+    {
+        $identity = $this->identityGenerator->generate();
+        $message = StringHelper::byteLength($identity) . '~' . $identity . '~' . $message;
+        $hash = hash_hmac($this->algorithm, $message, $this->secretKey, true);
+        if (!$hash) {
+            throw new RuntimeException("Failed to generate HMAC with hash algorithm: {$this->algorithm}.");
+        }
+        return $hash;
+    }
+
+    private function getHashLength(): int
+    {
+        return StringHelper::byteLength($this->generateHash(''));
     }
 }